random_scenario -> parametrized
