+++ task_generator/task_generator/constants.py
@@ -37,7 +37,7 @@ class Constants:
         RANDOM = "random"
         STAGED = "staged"
         SCENARIO = "scenario"
-        RANDOM_SCENARIO = "random_scenario"
+        PARAMETRIZED = "parametrized"
         DYNAMIC_MAP_RANDOM = "dynamic_map_random"
         DYNAMIC_MAP_STAGED = "dynamic_map_staged"
