+++ task_generator/task_generator/tasks/__init__.py
@@ -6,6 +6,6 @@ from .task_factory import TaskFactory # noqa
 from .random import RandomTask  # noqa
 from .scenario import ScenarioTask  # noqa
 from .staged import StagedRandomTask  # noqa
-from .random_scenario import RandomScenarioTask  # noqa
+from .parametrized import ParametrizedTask  # noqa
 from .dynamic_map_random import DynamicMapRandomTask  # noqa
 from .dynamic_map_staged import DynamicMapStagedRandomTask  # noqa
